+++ lib/rpi_gpio.dart
@@ -10,7 +10,7 @@ class RpiGpio extends Gpio {
   static bool _instantiatedGpio = false;
 
   final _polledInputs = <RpiGpioInput>[];
-  Duration _pollingFrequency = new Duration(milliseconds: 10);
+  Duration _pollingFrequency = Duration(milliseconds: 10);
   Timer _pollingTimer;
 
   /// Instantiate a new GPIO manager.
@@ -18,12 +18,11 @@ class RpiGpio extends Gpio {
   /// If you want to use the I2C pins as GPIO pins, pass `i2c: false`.
   /// If you want to use the SPI pins as GPIO pins, pass `spi: false`.
   RpiGpio({bool i2c = true, bool spi = true, bool eeprom = true}) {
-    if (_instantiatedGpio)
-      throw new GpioException('RpiGpio already instantiated');
+    if (_instantiatedGpio) throw GpioException('RpiGpio already instantiated');
     _instantiatedGpio = true;
     int result = _setupGpio();
     if (result != 0)
-      throw new GpioException('RpiGpio initialization failed: $result');
+      throw GpioException('RpiGpio initialization failed: $result');
     if (i2c) physI2CPins.forEach(allocatePin);
     if (spi) physSpiPins.forEach(allocatePin);
     if (eeprom) physEepromPins.forEach(allocatePin);
@@ -32,21 +31,21 @@ class RpiGpio extends Gpio {
   @override
   void throwPinAlreadyAllocated(int physicalPin) {
     if (physI2CPins.contains(physicalPin)) {
-      throw new GpioException(
+      throw GpioException(
           'If you want to use I2C pins for GPIO,'
           ' pass "i2c: false" into the RpiGpio constructor\n'
           'Already allocated',
           physicalPin);
     }
     if (physSpiPins.contains(physicalPin)) {
-      throw new GpioException(
+      throw GpioException(
           'If you want to use SPI pins for GPIO,'
           ' pass "spi: false" into the RpiGpio constructor\n'
           'Already allocated',
           physicalPin);
     }
     if (physEepromPins.contains(physicalPin)) {
-      throw new GpioException(
+      throw GpioException(
           'If you want to use EEPROM pins for GPIO,'
           ' pass "eeprom: false" into the RpiGpio constructor\n'
           'Already allocated',
@@ -58,20 +57,26 @@ class RpiGpio extends Gpio {
   @override
   void dispose() {
     int result = _disposeGpio();
-    if (result != 0) throw new GpioException('RpiGpio dispose failed: $result');
+    if (result != 0) throw GpioException('RpiGpio dispose failed: $result');
     _instantiatedGpio = false;
   }
 
   @override
   GpioInput input(int physicalPin, [Pull pull = Pull.off]) {
     allocatePin(physicalPin);
-    return new RpiGpioInput._(this, physicalPin, pull);
+    return RpiGpioInput._(this, physicalPin, pull);
   }
 
   @override
   GpioOutput output(int physicalPin) {
     allocatePin(physicalPin);
-    return new RpiGpioOutput._(physicalPin);
+    return RpiGpioOutput._(this, physicalPin);
+  }
+
+  @override
+  GpioPwm pwm(int physicalPin) {
+    allocatePin(physicalPin);
+    return RpiGpioPwm._(this, physicalPin);
   }
 
   @override
@@ -90,12 +95,12 @@ class RpiGpio extends Gpio {
 
   void _startPollingTimer() {
     if (_polledInputs.isNotEmpty && _pollingFrequency != null) {
-      _pollingTimer ??= new Timer.periodic(_pollingFrequency, _pollInputs);
+      _pollingTimer ??= Timer.periodic(_pollingFrequency, _pollInputs);
     }
   }
 
   void _pollInputs(_) {
-    final polledInputs = new List<RpiGpioInput>.from(_polledInputs);
+    final polledInputs = List<RpiGpioInput>.from(_polledInputs);
     for (RpiGpioInput input in polledInputs) {
       input._poll();
     }
@@ -115,17 +120,23 @@ class RpiGpio extends Gpio {
   int _disposeGpio() native "disposeGpio";
 }
 
-class RpiGpioInput extends GpioInput {
+class _RpiGpioCommon {
   final RpiGpio gpio;
   final int physicalPin;
   final int bcmGpio;
+
+  _RpiGpioCommon._(this.gpio, this.physicalPin)
+      : bcmGpio = physToBcmGpioRPi2[physicalPin];
+}
+
+class RpiGpioInput extends _RpiGpioCommon with GpioInput {
   final Pull pull;
 
   StreamController<bool> _valuesController;
   bool _lastValue;
 
-  RpiGpioInput._(this.gpio, this.physicalPin, this.pull)
-      : bcmGpio = physToBcmGpioRPi2[physicalPin] {
+  RpiGpioInput._(RpiGpio gpio, int physicalPin, this.pull)
+      : super._(gpio, physicalPin) {
     _setGpioInput(bcmGpio, pull.index);
   }
 
@@ -135,10 +146,10 @@ class RpiGpioInput extends GpioInput {
   @override
   Stream<bool> get values {
     if (_valuesController != null)
-      throw new GpioException(
+      throw GpioException(
           'cancel existing values stream before calling values again',
           physicalPin);
-    _valuesController = new StreamController(onListen: () {
+    _valuesController = StreamController(onListen: () {
       _poll();
       gpio._startPolling(this);
     }, onCancel: () {
@@ -161,20 +172,63 @@ class RpiGpioInput extends GpioInput {
   bool _readGpio(int bcmGpioPin) native "readGpio";
 }
 
-class RpiGpioOutput extends GpioOutput {
-  final int physicalPin;
-  final int bcmGpio;
+class _RpiGpioCommonOutput extends _RpiGpioCommon {
+  _RpiGpioCommonOutput._(RpiGpio gpio, int physicalPin) : super._(gpio, physicalPin) {
+    _setGpioOutput(bcmGpio);
+  }
 
-  RpiGpioOutput._(this.physicalPin) : bcmGpio = physToBcmGpioRPi2[physicalPin] {
+  void _setGpioOutput(int bcmGpio) native "setGpioOutput";
+  void _writeGpio(int bcmGpioPin, bool newValue) native "writeGpio";
+}
+
+class RpiGpioOutput extends _RpiGpioCommonOutput with GpioOutput {
+  RpiGpioOutput._(RpiGpio gpio, int physicalPin) : super._(gpio, physicalPin) {
     _setGpioOutput(bcmGpio);
   }
 
   @override
   set value(bool newValue) {
-    if (newValue == null) throw new GpioException('Invalid value', physicalPin);
+    if (newValue == null) throw GpioException('Invalid value', physicalPin);
     _writeGpio(bcmGpio, newValue);
   }
+}
 
-  void _setGpioOutput(int bcmGpio) native "setGpioOutput";
-  void _writeGpio(int bcmGpioPin, bool newValue) native "writeGpio";
+class RpiGpioPwm extends _RpiGpioCommonOutput with GpioPwm {
+  int _dutyCycle = 0;
+  Timer _currentTimer;
+
+  RpiGpioPwm._(RpiGpio gpio, int physicalPin) : super._(gpio, physicalPin);
+
+  @override
+  void set dutyCycle(int percentOn) {
+    if (percentOn <= 0) {
+      if (_dutyCycle == 0) return;
+      _dutyCycle = 0;
+      _writeGpio(bcmGpio, false);
+      _currentTimer?.cancel();
+      _currentTimer = null;
+      return;
+    } else if (percentOn >= 100) {
+      if (_dutyCycle == 100) return;
+      _dutyCycle = 100;
+      _writeGpio(bcmGpio, true);
+      _currentTimer?.cancel();
+      _currentTimer = null;
+      return;
+    }
+    if (_dutyCycle == percentOn) return;
+    _dutyCycle = percentOn;
+    if (_currentTimer == null) {
+      _currentTimer = Timer.periodic(const Duration(microseconds: 100), _pulse);
+      _pulse(_currentTimer);
+    }
+  }
+
+  void _pulse(Timer timer) async {
+    _writeGpio(bcmGpio, true);
+    int currentTick = timer.tick;
+    await Future.delayed(Duration(microseconds: _dutyCycle));
+    if (!timer.isActive || timer.tick != currentTick) return;
+    _writeGpio(bcmGpio, false);
+  }
 }