+++ test/debouncer_test.dart
@@ -7,13 +7,13 @@ import '../example/debouncer.dart';
 const testPinNum = 2;
 
 void main() {
-  var timeLimit = new Duration(milliseconds: 10);
+  var timeLimit = Duration(milliseconds: 10);
   Completer<bool> completer;
   Completer cancelCompleter;
-  var controller = new StreamController<bool>(onCancel: () {
+  var controller = StreamController<bool>(onCancel: () {
     cancelCompleter.complete();
   });
-  var debouncer = new Debouncer(false, 1);
+  var debouncer = Debouncer(false, 1);
   StreamSubscription subscription =
       controller.stream.transform(debouncer).listen((bool newValue) {
     completer.complete(newValue);
@@ -21,14 +21,14 @@ void main() {
   });
 
   test('one event', () async {
-    completer = new Completer<bool>();
+    completer = Completer<bool>();
     controller.add(true);
     var future = completer.future.timeout(timeLimit);
     expect(await future, true);
   });
 
   test('two events ignored', () async {
-    completer = new Completer<bool>();
+    completer = Completer<bool>();
     controller.add(false);
     controller.add(true);
     try {
@@ -40,18 +40,18 @@ void main() {
   });
 
   test('three events collapsed to one', () async {
-    completer = new Completer<bool>();
+    completer = Completer<bool>();
     controller.add(false);
     controller.add(true);
     controller.add(false);
     // Expect one result ... no exceptions from additional events
     expect(await completer.future.timeout(timeLimit), false);
-    await new Future.delayed(timeLimit);
+    await Future.delayed(timeLimit);
   });
 
   test('cancel', () async {
-    cancelCompleter = new Completer();
-    subscription.cancel();
+    cancelCompleter = Completer();
+    await subscription.cancel();
     await cancelCompleter.future.timeout(timeLimit);
   });
 }